Split 'UserAuthService' class functionality to 'UserAccessService' and 'UserAuthService'
